Add endpoint for playlist updates
Unlike series/video metadata and acl updating,
this uses a single endpoint for both (which is how
it is done in Opencast). That means that besides the
playlist id, each parameter is optional and will only
be updated when it is `Some`.

Calling this API will update the Playlist in Tobira's
Database without waiting for sync.