Add endpoint for playlist deletion
This utilizes the generic `delete_opencast_thing`
endpoint that is also used for series and events.
The corresponding playlist endpoint in Opencast
however works a little different: There are no
workflows or other magic background operations
involved. When we get a 200 response, it simply
means that the playlist has been deleted in Opencast.

This in turn lets us avoid having to create yet
another table/view that tracks "deleted" playlists
(which we have for series and events). We can
simply go ahead and delete the playlist directly
in our DB, no need for a `pending` limbo state and
no fear of sync concflicts.