Move playlist mutations to dedicated file
**Logic of mutations remains untouched**

Since `model/playlist` is a folder rather than a
single file and it already has a `mod.rs` file, I
think it's nicer to also have the mutations in a
separate file and not cram everything in `mod.rs`.
I don't know about best practices for this,
and `series` and `events` do do everything in one
file. I personally prefer separation of concerns
(at least in this case).