Add "playlist" item type
I didn't do this before because I wasn't
sure what to do about the playlist's state/status.
But as far as I can tell, playlists do not need
that, since their API works a little different.
Creating, updating and deleting will either succeed
or fail, and we get that result as an immediate
response from Opencast. So when it fails, we just
don't update it. This means that it should always
be in sync.