Add metadata editing to playlist details
This had to adjust the way metadata
is passed to the API. Instead of having
both title and description optional, it
now uses the `BasicMetadata` struct again,
which makes sure that the title won't be
overwritten with a `None` value when
omitted from the mutation. At the same time,
for better or worse, the whole `update_playlist`
API retains its swiss army knife character.