Add metadata editing to playlist details
This had to adjust the way metadata
is passed to the API. Instead of having
both title and description optional, it
now uses the `BasicMetadata` struct again,
which necessary for the frontend component
and where the title is required.

That component is used for both creating and
editing metadata, and while the title is
technically optional when editing, it is
mandatory for creation.