Add content editing to playlist details
This adds the interface to add and remove
videos to/from playlists. The backend and
API needed a slight adjustment, because
OC needs the `opencast_id` of the entries.
We could either pass them from frontend
(which would need some query adjustments)
or load them in backend (which is what I
do here now).