Handle unknown/missing videos in playlist editing
Previously they just weren't shown, and worse:
Any entries that are unknown to Tobira were
just dropped and lost upon using the `update`
endpoint.

This commit adds (1): a note to playlist editing
with the count of unknown videos, and
(2): some code in backend that adds these
entries back to the mutation.
Caveat: They will get new ids (the `entry_id`,
which is given by opencast, NOT the `opencast_id`,
that remains the same of course). But I don't think
that is an issue.