Add entries to playlist creation
It is now possible to set initial entries
when creating a playlist. I reused some
code from `EditVideoList` and factor out
some code to prevent duplication.
This has a small side effect: When adding
or removing videos with that component (before
submitting the form), the event count
above the interface will now update at the
same time and reflect the number of events
that is currently displayed. Previously, this
was using the count from the DB entry and
was only updated when the user submitted
the form.

Another thing: I wasn't sure whether to
integrate the events thing to react hook
form. State management is easier when it's
not integrated, and it is not necessary to
do any form validation on the entries field
(since that can also be left empty).
On the other hand, integration would be nice
at the very least for consistency.