Add thumbnail option to uploader
This will allow users to upload custom thumbnails
for their video. Since Tobira only supports single-stream
videos and their main type was previously chosen to be
"presentation", this continues to use "presentation" as
flavor for the thumbnail.
A vanilla Opencast instance uses preset workflows that
deal with uploaded thumbnails, and we piggyback on them
to keep things simple and make the thumbnail upload
possible without any prior configuration.

The following commit will make some things related to these
workflows configurable, but that is purely optional and
only needs to be done if the connected Opencast instance
uses workflows and/or values for these properties that
deviate from the default ones.