Adjust config docs
This adds some warning to the config docs to urge people
to still use `player+preview` as subtype flavor for their
final publications, as it was brought to my attention that this
subtype value is hardcoded in multiple places in Opencast and
thus is heavily relied upon.
Using other flavors for that just isn't supported.
That is only relevant when the *attachment* subtype is changed
to use something different from the default value.