Add second editor condition and adjust config docs
Turns out the editor needs `presenter/thumbnail_edited`
set in order to show the "presentation" thumbnail. I don't
know why and couldn't find out by looking through a lot of
Editor and Opencast code. As a workaround, this now sets both
`presentation` and `presenter` conditions to make sure the
uploaded thumbnail is used everywhere.

I also added some warning to the config docs to urge people
to still use `player+preview` as subtype flavor for their
final publications, as it was brought to my attention that this
subtype value is hardcoded in multiple places in Opencast  and
thus is heavily relied upon.
Using other flavors for that just isn't supported.
That is only relevant when the *attachment* subtype is changed
to use something different from the default value.