Add/modify `<itunes:author>` tag in rss feeds
Previously this was only adding the
author tag to the individual items
of an rss feed. I believe this was
an oversight/misunderstanding and
the tag was actually meant to be
added to the feed itself.

It was also requested that the tag
should use the `publisher` metadata
entry if available, which this commit
implements accordingly. That field isn't
mandatory so we do need one or more
fallback values:
- Playlists do not even
have a metadata field, so their feed
needs to always use the creator field
instead.
- Series will first look if the
`publisher` is specified and if it isn't,
checks whether there is a `creator` value
in the metadata.
- We do not harvest the `creator` field
for series in Tobira (might be something
we want to remedy), so if the Series'
metadata has neither `publisher` nor
`creator` specified, the author tag
is omitted
- Finally, the single items/episodes
have both metadata and creator(s) fields,
so we first look for `publisher` and
`creator` in that metadata and then fallback
to the creators field