Keep using `String` instead of `OpencastId` type in GraphQL queries
Unfortunately, GraphQL doesn't know that an OpencastId
is just a String in disguise, and we can't tell it
either. So to keep things backwards compatible with
Opencast (which also uses these `by_opencast_id`
queries), this has to stick with the explicit `String`
type.

Note on playlists: Opencast and the Admin UI don't have
any Playlist features yet that could talk with the
respective endpoint, but I expect this will come in the
future. While we could in theory say "use our special
type and deal with it", I think it's nicer if it can
mirror the existing series and event endpoints.