Add `manage` button to single series pages
This button not shown in a regular realm context,
but only in series blocks on single series pages.
That somewhat mirrors what we do with video blocks.
Putting it everywhere would clutter up the UI.

On the other hand, this might not achieve the desired
effect of making it easier to get to that manage page
in the first place. So maybe it should actually be shown
in every context, but only for people with write access.