Add `canWrite` check to series manage button
And also show that button on every block if the
conditions allow (see the other commit message for
reasoning).

Also removes that other condition (`isSingleItemPage`)
or rather the function that returns the condition.
It wasn't used anyway (guess that got lost in some edit
again.... gotta pay more attention when using `ctrl + z`).