Add embed url to share button
It was requested to supply an easy way
to get the embed url without having to
manually copy it from the iframe.

In order to not blow up the vertical
space needed by the menu, I shrunk the
<iframe> copy thingy a bit, which should
be fine. It was larger because that would
make it easier to get to the url in the first
place (I think). ~~It's still scrollable.~~

This is just an admittedly low-effort
proposition, but I'm not sure if there
is or needs to be a nicer way.