Treat newly created series as "synced" immediately
We already know that the series has been created
successfully when POST returns 201, so the series
might as well be marked as "ready" even though it's
not technically synced yet.
This way, users can immediately start adding videos
or change metadata if they goofed up when creating
the series.

We could technically send a GET request first to
use the actual factual data that's coming from
Opencast, but Tobira already knows title, (optional)
description and ACL, so there's little benefit in that.