Change `NotAuthorized` note from error to info
And make the login note a clickable link.

The red of the error note was perceived as too
alarmist and I agree to an extent. It's not really
an error. The grey that is used now is a littler darker
to provide enough contrast to other regular elements.
But with that, the login link had too little contrast
to meet a11y criteria, so I had to make that darker as well.
It's a little too dark now in my opinion. I might try a
`0.5` variant for the primary color.